+++ server/src/jobs/parcoursup/export/index.js
@@ -1,14 +1,9 @@
 const { runScript } = require("../../scriptWrapper");
 const logger = require("../../../common/logger");
 const { Formation } = require("../../../common/model/index");
-const { createParcoursupToken } = require("../../../common/utils/jwtUtils");
-const axios = require("axios");
+const parcoursupApi = require("./parcoursupApi");
 
-const privateKey = process.env.CATALOGUE_APPRENTISSAGE_PARCOURSUP_PRIVATE_KEY.replace(/\\n/gm, "\n");
-const pwd = process.env.CATALOGUE_APPRENTISSAGE_PARCOURSUP_PRIVATE_KEY_PWD;
-const id = process.env.CATALOGUE_APPRENTISSAGE_PARCOURSUP_CERTIFICATE_ID;
 const limit = Number(process.env.CATALOGUE_APPRENTISSAGE_PARCOURSUP_LIMIT || 50);
-const endpoint = process.env.CATALOGUE_APPRENTISSAGE_PARCOURSUP_ENDPOINT;
 
 const filter = {
   parcoursup_statut: "en attente de publication",
@@ -32,6 +27,12 @@ const select = {
   bcn_mefs_10: 1,
   rome_codes: 1,
   parcoursup_statut_history: 1,
+  parcoursup_error: 1,
+};
+
+// Retry the ones with errors last
+const sort = {
+  parcoursup_error: 1,
 };
 
 const formatter = ({ rncp_code, cfd, uai_formation, id_rco_formation, bcn_mefs_10 = [], rome_codes = [] }) => {
@@ -41,10 +42,37 @@ const formatter = ({ rncp_code, cfd, uai_formation, id_rco_formation, bcn_mefs_1
     rncp: Number(rncp_code.replace("RNCP", "")),
     cfd,
     uai: uai_formation,
-    rco: id_rco_formation,
+    rco: id_rco_formation, // TODO send cle_ministere_educatif
     mef,
     rome: rome_codes,
   };
+};
+
+const createCursor = () => {
+  return Formation.find(filter, select).sort(sort).limit(limit).cursor();
+};
+
+const createFormation = async (formation) => {
+  try {
+    const data = formatter(formation);
+    const responseData = await parcoursupApi.postFormation(data);
+
+    // console.log("Parcoursup WS response", responseData);
+    formation.parcoursup_id = responseData.g_ta_cod;
+    formation.parcoursup_statut = "publié";
+    formation.last_update_at = Date.now();
+    formation.last_update_who = "web service Parcoursup";
+    formation.parcoursup_statut_history.push({
+      date: new Date(),
+      parcoursup_statut: "publié",
+      last_update_who: "web service Parcoursup",
+    });
+    await formation.save();
+  } catch (e) {
+    logger.error("Parcoursup WS error", e?.response?.data ?? e);
+    formation.parcoursup_error = e?.response?.data ?? "erreur de création";
+    await formation.save();
+  }
 };
 
 /**
@@ -52,30 +80,9 @@ const formatter = ({ rncp_code, cfd, uai_formation, id_rco_formation, bcn_mefs_1
  * pour création sur Parcoursup via le Web Service dédié
  */
 const run = async () => {
-  let cursor = Formation.find(filter, select).limit(limit).cursor();
+  let cursor = createCursor();
   for await (const formation of cursor) {
-    const data = formatter(formation);
-    const token = createParcoursupToken({ data, privateKey, pwd, id });
-
-    try {
-      const { data: responseData } = await axios.post(endpoint, data, {
-        headers: { Authorization: `Bearer ${token}` },
-      });
-
-      // console.log("Parcoursup WS response", responseData);
-      formation.parcoursup_id = responseData.g_ta_cod;
-      formation.parcoursup_statut = "publié";
-      formation.last_update_at = Date.now();
-      formation.last_update_who = "web service Parcoursup";
-      formation.parcoursup_statut_history.push({
-        date: new Date(),
-        parcoursup_statut: "publié",
-        last_update_who: "web service Parcoursup",
-      });
-      await formation.save();
-    } catch (e) {
-      logger.error("Parcoursup WS error", e?.response?.data ?? e);
-    }
+    await createFormation(formation);
   }
 };
 
@@ -88,3 +95,8 @@ if (process.env.standalone) {
     logger.info(" -- End psup export -- ");
   });
 }
+
+module.exports = {
+  createCursor,
+  createFormation,
+};
